Config sequelize
