+++ package.json
@@ -12,8 +12,10 @@
     "cors": "^2.8.5",
     "dotenv": "^9.0.1",
     "express": "^4.17.1",
-    "knex": "^0.95.9",
-    "pg": "^8.7.1"
+    "pg": "^8.7.1",
+    "pg-hstore": "^2.3.4",
+    "sequelize": "^6.6.5",
+    "sequelize-cli": "^6.2.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.7"
