--- src/database/connect.js
@@ -1,5 +1,0 @@
-const environment = process.env.NODE_ENV || 'development';
-const config = require('../../knexfile')[environment];
-const knex = require('knex')(config);
-
-module.exports = knex;
