--- src/database/seeds/01_municipios.js
@@ -1,14 +1,0 @@
-
-exports.seed = function(knex) {
-  // Deletes ALL existing entries
-  return knex('municipios').del()
-    .then(function () {
-      // Inserts seed entries
-      return knex('municipios').insert([
-        { 
-          id_municipio: 1,
-          nome_municipio: "Salvaterra" 
-        },
-      ]);
-    });
-};
