--- src/index.js
@@ -1,6 +1,0 @@
-const app = require('./server');
-const port = process.env.PORT || 3000;
-
-app.listen(port, () => {
-	console.log(`http://localhost:${port}`);
-});
