+++ src/server.js
@@ -1,18 +1,6 @@
-const express = require('express');
-const cors = require('cors');
+const app = require('./app');
+const port = process.env.PORT || 3000;
 
-const routesMunicipios = require('./routes/routesMunicipios');
-const routesPraias = require('./routes/routesPraias');
-const routesAvaliacao = require('./routes/routesAvaliacao');
-
-const app = express();
-
-app.use(cors());
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
-app.use(routesMunicipios);
-app.use(routesPraias);
-app.use(routesAvaliacao);
-
-module.exports = app;
+app.listen(port, () => {
+	console.log(`http://localhost:${port}`);
+});
